+++ lib_fiber/cpp/include/fiber/fiber_cond.hpp
@@ -24,7 +24,7 @@ public:
 	 * @param timeout {int} 超时等待时间（毫秒）
 	 * @return {bool} 成功时返回 true，否则返回 false 表示超时
 	 */
-	bool wait(fiber_event& event, long long timeout = -1);
+	bool wait(fiber_event& event, int timeout = -1);
 
 	/**
 	 * 唤醒在条件变量上的等待者，如果没有等待者则直接返回，运行行为和
